// src/Shared/scripts/Enums/20 - MenuCommand.ts
export enum MenuCommandKey {
    Unknown = 0,
    AddCeTab = 20360207,
    Cancel = 20784806,
    CloseWindow = 20366927,
    CompactScUi = 20253876,
    DebugClearPopUpConsole = 20984733,
    DebugForceAutoSnapShot = 20906196,
    DebugTriggerPopUpReload = 20975662,
    DrawStorage = 20084759,
    DrawStoragePopUpLogStorage = 20648858,
    GoContentEditor = 20730862,
    GoDesktop = 20049771,
    GoScModeEdit = 20795022,
    Ping = 20501215,
    PresentationDetails = 20721120,
    PutAdminB = 20104667,
    QuickPublish = 20239875,
    Remove = 20395539,
    RestoreStateNewTab = 20620284,
    RestoreStateSameTab = 20042016,
    RestoreStateTBDTab = 20013232,
    ScModeNormal = 20979923,
    ScModePrev = 20837972,
    TakeSnapShot = 20006262,
    ToggleFavorite = 20625140,
    UpdateNickname = 20900113,
    DebugTriggerfatalError
}